pyproject.toml
[project]

name = "gage"
version = "0.1.0-dev.0"
authors = [
  { name="Garrett Smith", email="garrett@placeon.earth" },
]
description = "Model evaluation and publishing toolkit"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "click==8.1.7",
    "chardet==5.2.0",
    "human-readable==1.3.4",
    "jschon==0.11.0",
    "lark==1.1.7",
    "libcst==1.0.1",
    "natsort==8.4.0",
    "proquint==0.2.1",
    "psutil==5.9.5",
    "python-ulid==2.2.0",
    "pyyaml==6.0.1",
    "rich==13.5.2",
    "setuptools==68.2.2",
    "tomli==2.0.1",
    "typer@git+https://github.com/gar1t/typer",
]

[project.optional-dependencies]

dev = [
    "black",
    "groktest",
]

[tool.setuptools.package-data]

gage = [
    "*.schema.json"
]

[project.scripts]

gage = "gage.__main__:main"

[project.urls]

"Homepage" = "https://github.com/gageml/gage"
"Bug Tracker" = "https://github.com/gageml/gage/issues"

[tool.black]

skip-string-normalization = true
extend-exclude = "/tests/|/gage/_vendor/"

[tool.groktest]

include = "tests/*.md"
python.init = "from gage._internal.test import *"
parse.functions = "gage._internal.test"
